# docs/game-runtime-openapi.yaml
openapi: 3.1.0
info:
  title: Vifu Game Runtime API
  version: 1.0.0
  description: Engine-neutral API for published Vifu Game releases and durable sessions.
servers:
  - url: http://localhost:6790/{projectSlug}/v1/game
    variables:
      projectSlug:
        default: my-game
security:
  - bearerAuth: []
paths:
...
  /run:
    post:
      operationId: runGame
      summary: Create a durable session and submit game.start
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunGameRequest"
      responses:
        "201":
          description: Session and first stable runtime boundary
          content:
            application/json:
              schema:
                type: object
                required: [sessionId, releaseId, advance, optionalCapabilityFallbacks]
                properties:
                  sessionId:
                    type: string
                    format: uuid
                  releaseId:
                    type: string
                    format: uuid
                  advance:
                    $ref: "#/components/schemas/PublicAdvance"
                  optionalCapabilityFallbacks:
                    type: array
                    items:
                      type: string
        default:
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Project API key with Game Runtime Execute permission.
...
  responses:
    Error:
      description: JSON error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    HostDescriptor:
      type: object
      required: [engine, capabilities]
      additionalProperties: false
      properties:
        engine:
          type: string
          minLength: 1
        adapterVersion:
          type: string
        capabilities:
          type: array
          items:
            type: string
        locale:
          type: string
        bindingManifestHash:
          type: string
...
    RunGameRequest:
      type: object
      required: [host]
      additionalProperties: false
      properties:
        host:
          $ref: "#/components/schemas/HostDescriptor"
        input: {}
        idempotencyKey:
          type: string
          minLength: 1
          maxLength: 200
...
    PublicAdvance:
      type: object
      required: [status, revision, outstandingHostActions, events]
      properties:
        status:
          $ref: "#/components/schemas/SessionStatus"
        revision:
          type: integer
          minimum: 0
        publicOutput: {}
        outstandingHostActions:
          type: array
          items:
            $ref: "#/components/schemas/PendingHostAction"
        failure:
          $ref: "#/components/schemas/RuntimeFailure"
        events:
          type: array
          items:
            $ref: "#/components/schemas/GameEvent"
    PendingHostAction:
      type: object
      required: [actionId, nodeId, target, action, arguments]
      properties:
        actionId:
          type: string
        nodeId:
          type: string
        target:
          type: string
        action:
          type: string
        arguments: {}
    GameEvent:
      type: object
      required: [specversion, id, source, type, sequence, data]
      properties:
        specversion:
          type: string
          const: "1.0"
        id:
          type: string
        source:
          type: string
        type:
          type: string
        subject:
          type: string
        sequence:
          type: integer
          minimum: 1
        data: {}
    SessionStatus:
      type: string
      enum:
        - running
        - waiting_input
        - waiting_effect
        - waiting_host
        - completed
        - failed
        - cancelled
    RuntimeFailure:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        nodeId:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string